+++ packages/orq-rc/RELEASES.md
@@ -9179,3 +9179,13 @@ Based on:
 - [typescript v4.10.0-rc.3] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.3] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.3 - packages/orq-rc
+
+## 2026-05-14 10:26:37
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.0 (2.884.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.4] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.4] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.4 - packages/orq-rc